add yes or no question

+++ googlectl/delete.py
@@ -18,7 +18,9 @@ class UserDelete(Command):
 
     def take_action(self, parsed_args):
         client = googlectl.libgooglectl.Client()
-        client.delete_user(parsed_args.email)
+        response = client.yes_or_no("Are you sure to delete {} ?".format(parsed_args.email))
+        if response:
+            client.delete_user(parsed_args.email)
 
 class GroupDelete(Command):
     "Delete the group"
@@ -32,7 +34,9 @@ class GroupDelete(Command):
 
     def take_action(self, parsed_args):
         client = googlectl.libgooglectl.Client()
-        client.delete_group(parsed_args.groupkey)
+        response = client.yes_or_no("Are you sure to delete {} ?".format(parsed_args.groupkey))
+        if response:
+            client.delete_group(parsed_args.groupkey)
 
 
 class GroupMemberDelete(Command):
@@ -48,4 +52,7 @@ class GroupMemberDelete(Command):
 
     def take_action(self, parsed_args):
         client = googlectl.libgooglectl.Client()
-        client.delete_member(parsed_args.groupkey, parsed_args.email)
+        response = client.yes_or_no("Are you sure to delete {} from {} ?".format(parsed_args.email, parsed_args.groupkey))
+
+        if response:
+            client.delete_member(parsed_args.groupkey, parsed_args.email)
